+++ buildSrc/src/main/kotlin/Releases.kt
@@ -72,10 +72,10 @@ object Releases {
     }
   }
 
-  object ImplementationGuide : LibraryArtifact {
-    override val artifactId = "implementationguide"
+  object Knowledge : LibraryArtifact {
+    override val artifactId = "knowledger"
     override val version = "0.1.0-alpha001"
-    override val name = "Android FHIR Implementation Guide Library"
+    override val name = "Android FHIR Knowledge Manager Library"
   }
 
   // Demo apps
