--- implementationguide/build.gradle.kts
@@ -1,131 +1,0 @@
-import java.net.URL
-
-plugins {
-  id(Plugins.BuildPlugins.androidLib)
-  id(Plugins.BuildPlugins.kotlinAndroid)
-  id(Plugins.BuildPlugins.kotlinKapt)
-  id(Plugins.BuildPlugins.mavenPublish)
-  jacoco
-  id(Plugins.BuildPlugins.dokka).version(Plugins.Versions.dokka)
-}
-
-publishArtifact(Releases.ImplementationGuide)
-
-createJacocoTestReportTask()
-
-android {
-  compileSdk = Sdk.compileSdk
-
-  defaultConfig {
-    minSdk = Sdk.minSdk
-    targetSdk = Sdk.targetSdk
-    testInstrumentationRunner = Dependencies.androidJunitRunner
-    // Need to specify this to prevent junit runner from going deep into our dependencies
-    testInstrumentationRunnerArguments["package"] = "com.google.android.fhir.implementationguide"
-  }
-
-  sourceSets {
-    getByName("androidTest").apply { resources.setSrcDirs(listOf("testdata")) }
-
-    getByName("test").apply { resources.setSrcDirs(listOf("testdata")) }
-  }
-
-  buildTypes {
-    getByName("release") {
-      isMinifyEnabled = false
-      proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
-    }
-  }
-
-  compileOptions {
-    isCoreLibraryDesugaringEnabled = true
-    sourceCompatibility = Java.sourceCompatibility
-    targetCompatibility = Java.targetCompatibility
-  }
-
-  packagingOptions {
-    resources.excludes.addAll(
-      listOf(
-        "license.html",
-        "META-INF/ASL2.0",
-        "META-INF/ASL-2.0.txt",
-        "META-INF/DEPENDENCIES",
-        "META-INF/LGPL-3.0.txt",
-        "META-INF/LICENSE",
-        "META-INF/LICENSE.txt",
-        "META-INF/license.txt",
-        "META-INF/license.html",
-        "META-INF/LICENSE.md",
-        "META-INF/NOTICE",
-        "META-INF/NOTICE.txt",
-        "META-INF/NOTICE.md",
-        "META-INF/notice.txt",
-        "META-INF/LGPL-3.0.txt",
-        "META-INF/sun-jaxb.episode",
-      )
-    )
-  }
-
-  kotlinOptions { jvmTarget = Java.kotlinJvmTarget.toString() }
-
-  configureJacocoTestOptions()
-}
-
-configurations {
-  all {
-    exclude(module = "xpp3")
-    exclude(module = "xpp3_min")
-  }
-}
-
-dependencies {
-  androidTestImplementation(Dependencies.AndroidxTest.core)
-  androidTestImplementation(Dependencies.AndroidxTest.runner)
-  androidTestImplementation(Dependencies.AndroidxTest.extJunitKtx)
-  androidTestImplementation(Dependencies.Kotlin.kotlinCoroutinesTest)
-  androidTestImplementation(Dependencies.junit)
-  androidTestImplementation(Dependencies.truth)
-
-  api(Dependencies.HapiFhir.structuresR4) { exclude(module = "junit") }
-
-  coreLibraryDesugaring(Dependencies.desugarJdkLibs)
-
-  implementation(Dependencies.Kotlin.stdlib)
-  implementation(Dependencies.Lifecycle.liveDataKtx)
-  implementation(Dependencies.Room.ktx)
-  implementation(Dependencies.Room.runtime)
-  implementation(Dependencies.timber)
-
-  kapt(Dependencies.Room.compiler)
-
-  testImplementation(Dependencies.AndroidxTest.archCore)
-  testImplementation(Dependencies.AndroidxTest.core)
-  testImplementation(Dependencies.junit)
-  testImplementation(Dependencies.Kotlin.kotlinCoroutinesTest)
-  testImplementation(Dependencies.mockitoInline)
-  testImplementation(Dependencies.mockitoKotlin)
-  testImplementation(Dependencies.robolectric)
-  testImplementation(Dependencies.truth)
-}
-
-tasks.dokkaHtml.configure {
-  outputDirectory.set(
-    file(
-      "../docs/${Releases.ImplementationGuide.artifactId}/${Releases.ImplementationGuide.version}"
-    )
-  )
-  suppressInheritedMembers.set(true)
-  dokkaSourceSets {
-    named("main") {
-      moduleName.set(Releases.ImplementationGuide.artifactId)
-      moduleVersion.set(Releases.ImplementationGuide.version)
-      noAndroidSdkLink.set(false)
-      externalDocumentationLink {
-        url.set(URL("https://hapifhir.io/hapi-fhir/apidocs/hapi-fhir-structures-r4/"))
-        packageListUrl.set(
-          URL("https://hapifhir.io/hapi-fhir/apidocs/hapi-fhir-structures-r4/element-list")
-        )
-      }
-    }
-  }
-}
