--- implementationguide/src/main/AndroidManifest.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<manifest package="com.google.android.fhir.implementationguide">
-
-    <application />
-
-    </manifest>
