+++ knowledge/src/main/AndroidManifest.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<manifest package="com.google.android.fhir.knowledge">
+
+    <application />
+
+    </manifest>
